pyproject.toml
[project]
name = "audio-text-tests"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.0",
    "cython>=3.2.1",
    "datasets>=4.4.1",
    "eng-to-ipa>=0.0.2",
    "evaluate>=0.4.6",
    "fire>=0.7.1",
    "gradio>=6.0.1",
    "hdbscan>=0.8.40",
    "inflect>=7.5.0",
    "jieba>=0.42.1",
    "jiwer>=4.0.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.7",
    "mutagen>=1.47.0",
    "numba>=0.62.1",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pydub>=0.25.1",
    "pypinyin>=0.55.0",
    "python-gitlab>=7.1.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "resemblyzer>=0.1.4",
    "scikit-learn>=1.7.2",
    "sentence-transmorgrifier>=0.0.1",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchcodec>=0.8.1",
    "torchdiffeq>=0.2.5",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "unidecode>=1.4.0",
    "uroman>=1.3.1.1",
]

[tool.pylint.messages_control]
disable = [
    "broad-exception-caught",
    "line-too-long"
]
